debugging

+++ internal/types.go
@@ -153,7 +153,7 @@ type HandlerFuncParams struct {
 	// GetObjectFrequency retrieves the access frequency count of a key. Can only be used with LFU type eviction policies.
 	GetObjectFrequency func(ctx context.Context, keys string) (int, error)
 	// GetObjectIdleTime retrieves the time in seconds since the last access of a key. Can only be used with LRU type eviction policies.
-	GetObjectIdleTime func(ctx context.Context, keys string) (int, error)
+	GetObjectIdleTime func(ctx context.Context, keys string) (float64, error)
 }
 
 // HandlerFunc is a functions described by a command where the bulk of the command handling is done.
